Fail more gracefully and use a lock when concurrent=1
Be very verbose about killing the child process in various edge cases.
Die if connection to zookeeper is lost, but continue to run if the
connection is suspended, when the connection is suspended, it is likely
zookeeper is restarting.

Also use a lock instead of a semaphore when there is only a concurrent
count of 1.